add freshexport

+++ config/models.py
@@ -12,8 +12,8 @@ class ElasticModel(models.Model):
 class FreshServiceModel(models.Model):
     service_url = models.URLField()
     api_key = models.CharField(max_length=200)
-    group_id = models.IntegerField()
-    requester_id = models.IntegerField()
+    group_id = models.BigIntegerField()
+    requester_id = models.BigIntegerField()
     requester_email = models.EmailField()
     requester_phone = models.CharField(max_length=200)
     ansprechperson = models.CharField(max_length=200)
